Moved consumer logic to UseCase.
