+++ RentIt.Bookings/RentIt.Bookings.Application/Interfaces/UseCases/Bookings/IUpdateBookingsAfterHousingChangedUseCase.cs
@@ -1,0 +1,9 @@
+﻿using RentIt.MessageBroker.Contracts.Events;
+
+namespace RentIt.Bookings.Application.Interfaces.UseCases.Bookings
+{
+    public interface IUpdateBookingsAfterHousingChangedUseCase
+    {
+        Task ExecuteAsync(HousingUpdatedEvent message, CancellationToken cancellationToken);
+    }
+}
